Split .bashrc setup like on Fedora
It seems that the structure of .bashrc is quite different on Fedora:
it has not provision for skipping setup on non-interactive shelss,
and pulls in /etc/bashrc. And it has support for ~/.bashrc.d/.
So let's move things into files under ~/.bashrc.d/, keep original
.bashrc on Fedora, and add support for ~/.bashrc.d/ into the
version of .bashrc that is installed on Debian/Ubuntu.

[I think the approach of modifying $PATH in .bashrc is broken:
we would need to do this for every shell and for non-shell
environments. This should be a systemd environment generator
instead.]