Switching to ansible roles.
The main idea behind this is to have roles that are more or less
self-contained, which are then included in the main playbook in all.yml
(or site.yml in the ansible docs).

The most specific ASPP settings (passwords/time zones/keyboard
layouts/wallpapers) should ideally not be inside the roles folders but
added to group_vars/host_vars directories (currently not used yet). This
should make it easier to find the changes necessary to set up a new
edition of ASPP.

Eventually, we probably want to move the list of very ASPP-specific
Python packages also to group_vars (or at least to a proper role var)
but this is not implemented yet.